requirements.txt
openai-whisper
streamlit
audio-recorder-streamlit
librosa
